LayoutTests/storage/indexeddb/removed-expected.txt

Ensure that some obsolete IndexedDB features are gone.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "removed.html"
PASS self.webkitIDBDatabaseError is undefined.
PASS 'IDBDatabaseException' in self is false
PASS 'errorCode' in indexedDB.open(dbname) is false
PASS 'setVersion' in IDBDatabase.prototype is false
PASS document.createEvent('IDBUpgradeNeededEvent') threw exception NotSupportedError: Failed to execute 'createEvent' on 'Document': The provided event type ('IDBUpgradeNeededEvent') is invalid..
PASS 'version' in document.createEvent('IDBVersionChangeEvent') is false
PASS successfullyParsed is true

TEST COMPLETE
